Deleted old version with simple scaling model
