+++ injector_model/__init__.py
@@ -1,5 +1,4 @@
 from .injector_model import InjectorChain
-from .injector_model_full_sc_integral import InjectorChain_v2
 from .space_charge_and_ibs import IBS_Growth_Rates, SC_Tune_Shifts
 from .parameters_and_helpers import Reference_Values, BeamParams_LEIR, BeamParams_PS, BeamParams_SPS
 from .sequence_makers import Sequences
