+++ injector_model/injector_model.py
@@ -1,79 +1,135 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 Simulation model of the CERN Injector Chain for different ions
+solving for full space charge (SC) lattice integral 
 - by Elias Waagaard 
 """
+import os
+from pathlib import Path
 import pandas as pd
 import numpy as np
-from pathlib import Path
 import math
-#from scipy.constants import e
 from scipy import constants
+import xtrack as xt
+import xpart as xp
+import matplotlib.pyplot as plt
 from collections import defaultdict
-import os
+
+from .parameters_and_helpers import Reference_Values, BeamParams_LEIR, BeamParams_PS, BeamParams_SPS
+from .sequence_makers import Sequences
+from .space_charge_and_ibs import SC_Tune_Shifts, IBS_Growth_Rates
+
+#### PLOT SETTINGS #######
+plt.rcParams.update(
+    {
+        "font.family": "serif",
+        "font.size": 20,
+        "axes.titlesize": 20,
+        "axes.labelsize": 18,
+        "xtick.labelsize": 18,
+        "ytick.labelsize": 18,
+        "legend.fontsize": 15,
+        "figure.titlesize": 20,
+    }
+)
 
 # Calculate the absolute path to the data folder relative to the module's location
 data_folder = Path(__file__).resolve().parent.joinpath('../data').absolute()
 energy_folder = Path(__file__).resolve().parent.joinpath('injection_energies/calculated_injection_energies').absolute()
+output_folder = Path(__file__).resolve().parent.joinpath('../output').absolute()
+
 
 class InjectorChain:
     """
-    Representation of the CERN Injector Chain for different ions with linear space charge effects 
-    to calculte maximum intensity limits
-    following Roderik Bruce's example from 2021
+    Representation of the CERN Injector Chain for different ions with full space charge lattice integral. 
+    This model accounts for
+    - full space charge integrals in LEIR, PS and SPS 
+
+    Parameters:
+    -----------
+    nPulsesLEIR : int
+        number of pulses injected into LEIR (maximum 7 today with Pb). If "None", then calculate max from SC limit
+    LEIR_bunches : int
+        after RF capture, how many bunches that will circulate in LEIR
+    PS_splitting : int
+        number of bunches that injected bunch in PS will be split into. E.g. "2" means that bunch will be split into 2 bunches
+    account_for_LEIR_ecooling : bool
+        whether to factor electron cooling time in LEIR, which is longer for lighter ions
+    LEIR_PS_strip : bool
+        whether stripping foil should be placed between LEIR and PS. If "False", then default stripping between PS and SPS
     """
-    def __init__(self, 
-                 ion_type = 'Pb',
-                 ion_type_ref='Pb',
-                 nPulsesLEIR = 1,
+    def __init__(self, ion_type='Pb', 
+                 nPulsesLEIR = None,
                  LEIR_bunches = 2,
                  PS_splitting = 2,
-                 account_for_SPS_transmission=True,
-                 LEIR_PS_strip=False,
-                 consider_PS_space_charge_limit=True,
-                 use_gammas_ref=False,
-                 higher_brho_LEIR=False,
-                 save_path_csv = 'output/csv_tables'
+                 account_for_LEIR_ecooling=False,
+                 LEIR_PS_strip=False
                  ):
         
+        # Import reference data and initiate ion
         self.full_ion_data = pd.read_csv("{}/Ion_species.csv".format(data_folder), header=0, index_col=0).T
         self.LEIR_PS_strip = LEIR_PS_strip
-        self.higher_brho_LEIR = higher_brho_LEIR
-        self.brho_string = '_higher_brho_LEIR' if self.higher_brho_LEIR else ''
-
-        # Check whether to load relativistic gamma data from injection_energies
-        self.use_gammas_ref = use_gammas_ref
-        
+        self.account_for_LEIR_ecooling = account_for_LEIR_ecooling
         self.init_ion(ion_type)
-        self.debug_mode = False
-        self.account_for_SPS_transmission = account_for_SPS_transmission
-        self.consider_PS_space_charge_limit = consider_PS_space_charge_limit
-        
-        # Initiate values for PS B-field
-        self.PS_MinB    = 383 * 1e-4 # [T] - minimum magnetic field in PS, (Gauss to Tesla) from Heiko Damerau
-        self.PS_MaxB    = 1.26 # [T] - minimum magnetic field in PS, from reyes Alemany Fernandez
-        self.PS_rho     = 70.079 # [m] - PS bending radius 
-        
+        self.load_Pb_lines()
+                
         # Rules for splitting and bunches 
         self.nPulsesLEIR = nPulsesLEIR
         self.LEIR_bunches = LEIR_bunches
         self.PS_splitting = PS_splitting
-        
-        # Also initiate reference values
-        self.ion_type_ref = ion_type_ref
-        self.ion0_referenceValues() 
-
-        # Save path
-        self.save_path = save_path_csv
-
-        #For printing some reference values
-        if self.debug_mode:
-            print(f"Initiate. Type: {self.ion_type}")
-            print("Q = {}, Z = {}\n".format(self.Q, self.Z))
 
 
-    def init_ion(self, ion_type):
+    def load_Pb_lines(self) -> None:
+        """
+        Load xtrack lines for LEIR, PS and SPS - retrieve Twiss tables and lengths
+        Calculate present space charge tune shift dQx and dQy for Pb, as a reference
+        """
+        # Instantiate reference values and class for tune shifts
+        ref_val = Reference_Values()
+        sc = SC_Tune_Shifts()
+
+        ##### LEIR #####
+        self.line_LEIR_Pb0 = Sequences.get_LEIR_line(m0_GeV = ref_val.m0_GeV, 
+                                            Q = ref_val.Q0_LEIR, 
+                                            gamma = ref_val.gamma0_LEIR_inj)
+        self.twiss_LEIR_Pb0 = self.line_LEIR_Pb0.twiss()
+        self.line_LEIR_length = self.line_LEIR_Pb0.get_length()
+        self.dQx0_LEIR_Pb, self.dQy0_LEIR_Pb = sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_LEIR_Pb0,
+                                                    particle_ref=self.line_LEIR_Pb0.particle_ref,
+                                                    line_length=self.line_LEIR_length,
+                                                    Nb=BeamParams_LEIR.Nb,
+                                                    beamParams=BeamParams_LEIR)
+
+        ##### PS #####
+        self.line_PS_Pb0 = Sequences.get_PS_line(m0_GeV = ref_val.m0_GeV, 
+                                            Q = ref_val.Q0_PS, 
+                                            gamma = ref_val.gamma0_PS_inj)
+        self.twiss_PS_Pb0 = self.line_PS_Pb0.twiss()
+        self.line_PS_length = self.line_PS_Pb0.get_length()
+        self.dQx0_PS_Pb, self.dQy0_PS_Pb = sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_PS_Pb0,
+                                                    particle_ref=self.line_PS_Pb0.particle_ref,
+                                                    line_length=self.line_PS_length,
+                                                    Nb=BeamParams_PS.Nb,
+                                                    beamParams=BeamParams_PS)
+
+        ##### SPS #####
+        self.line_SPS_Pb0 = Sequences.get_SPS_line(m0_GeV = ref_val.m0_GeV, 
+                                            Q = ref_val.Q0_SPS, 
+                                            gamma = ref_val.gamma0_SPS_inj)
+        self.twiss_SPS_Pb0 = self.line_SPS_Pb0.twiss()
+        self.line_SPS_length = self.line_SPS_Pb0.get_length()
+        self.dQx0_SPS_Pb, self.dQy0_SPS_Pb = sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_SPS_Pb0,
+                                                    particle_ref=self.line_SPS_Pb0.particle_ref,
+                                                    line_length=self.line_SPS_length,
+                                                    Nb=BeamParams_SPS.Nb,
+                                                    beamParams=BeamParams_SPS)
+        print('Pb tune shifts:')
+        print('LEIR: dQx_Pb = {:.3f}, dQy_Pb = {:.3f}'.format(self.dQx0_LEIR_Pb, self.dQy0_LEIR_Pb))
+        print('PS:   dQx_Pb = {:.3f}, dQy_Pb = {:.3f}'.format(self.dQx0_PS_Pb, self.dQy0_PS_Pb))
+        print('SPS:  dQx_Pb = {:.3f}, dQy_Pb = {:.3f}'.format(self.dQx0_SPS_Pb, self.dQy0_SPS_Pb))
+        print('Loaded all sequences.\n')
+
+
+    def init_ion(self, ion_type) -> None:
         """
         Initialize ion species for a given type 
         """
@@ -83,31 +139,33 @@ class InjectorChain:
         self.mass_GeV = self.ion_data['mass [GeV]']
         self.Z = self.ion_data['Z']
         self.A = self.ion_data['A']
-        self.Q = self.ion_data['Q before stripping']
+        
+        # Set charge state for diffent accelerators
+        self.Q_LEIR = self.ion_data['Q before stripping']
+        self.Q_PS = self.ion_data['Z'] if self.LEIR_PS_strip else self.ion_data['Q before stripping']
+        self.Q_SPS = self.ion_data['Z']
         
         # Values from first tables in Roderik's notebook
         self.linac3_current = self.ion_data['Linac3 current [uA]'] * 1e-6
         self.linac3_pulseLength = self.ion_data['Linac3 pulse length [us]'] * 1e-6
-
-        # General rules - stripping and transmission
-        self.LEIR_injection_efficiency = 0.5
-        self.LEIR_transmission = 0.8
         self.LEIR_PS_stripping_efficiency = self.ion_data['LEIR-PS Stripping Efficiency']
-        self.PS_transmission = 0.9
-        self.PS_SPS_transmission_efficiency = 1.0 # 0.9 is what we see today, but Roderik uses 1.0
-        self.PS_SPS_strip = not self.LEIR_PS_strip  # if we have LEIR-PS stripping, no stripping PS-SPS
-        self.PS_SPS_stripping_efficiency = 0.9  # default value until we have other value
-        self.SPS_transmission = 0.62
-        self.SPS_slipstacking_transmission = 1.0
+        self.load_ion_energy()
         
-        # Use gammas from calculated reference module
-        if self.use_gammas_ref:
-            self.load_ion_energy()
+        print(f"Initiated ion type: {self.ion_type}")
+        print("Q_LEIR = {}, Q_PS = {}, Q_SPS = {} (fully stripped)\nStrip LEIR-PS: {}".format(self.Q_LEIR, 
+                                                                                              self.Q_PS, 
+                                                                                              self.Q_SPS,
+                                                                                              self.LEIR_PS_strip))
+        
+        # LEIR ecooling rations
+        if self.account_for_LEIR_ecooling:
+            self.relative_ecooling_time_leir = self.ion_data['Relative_LEIR_ecooling_time']
+            print('E-cooling time w.r.t to Pb: {:.2f}\n'.format(self.relative_ecooling_time_leir))
 
 
     def load_ion_energy(self):
         """
-        Loads calculated ion energies for each ion type from the ion_injection_energies module
+        Loads calculated ion energies for all stable ion isotopes from the ion_injection_energies module
         """
         # Load ion energy data depending on where stripping is made 
         if self.LEIR_PS_strip:
@@ -115,17 +173,19 @@ class InjectorChain:
         else:
             self.ion_energy_data = pd.read_csv('{}/ion_injection_energies_PS_SPS_strip.csv'.format(energy_folder), index_col=0)
         
-        key = str(int(self.Q)) + self.ion_str + str(int(self.A))
+        # Load correct entry in ion energy table - key composed of lower charge state Q before stripping, then ion string, then mass number A
+        key = str(int(self.Q_LEIR)) + self.ion_str + str(int(self.A))
         ion_energy = self.ion_energy_data.loc[key]
         
         # Load reference injection energies
-        self.LEIR_gamma_inj_ref = ion_energy['LEIR_gamma_inj']
-        self.LEIR_gamma_extr_ref = ion_energy['LEIR_gamma_extr']
-        self.PS_gamma_inj_ref = ion_energy['PS_gamma_inj']
-        self.PS_gamma_extr_ref = ion_energy['PS_gamma_extr']
-        self.SPS_gamma_inj_ref = ion_energy['SPS_gamma_inj']
-        self.SPS_gamma_extr_ref = ion_energy['SPS_gamma_extr']
-        print('Ion type: {} and PS gamma: {}'.format(self.ion_type, self.PS_gamma_extr_ref))
+        self.LEIR_gamma_inj = ion_energy['LEIR_gamma_inj']
+        self.LEIR_gamma_extr = ion_energy['LEIR_gamma_extr']
+        self.PS_gamma_inj = ion_energy['PS_gamma_inj']
+        self.PS_gamma_extr = ion_energy['PS_gamma_extr']
+        self.SPS_gamma_inj = ion_energy['SPS_gamma_inj']
+        self.SPS_gamma_extr = ion_energy['SPS_gamma_extr']
+        print('\nIon type {}: \nPS extr gamma: {:.3f}'.format(self.ion_type, self.PS_gamma_extr))
+
 
     def beta(self, gamma):
         """
@@ -152,363 +212,295 @@ class InjectorChain:
         """
         Brho = p / (q * constants.c) # in Tm
         return Brho    
-    
-    
-    def spaceChargeScalingFactor(self, Nb, m, gamma, epsilon, sigma_z, fully_stripped=True):
+       
+
+    def LEIR_SC_limit(self):
         """
-        Approximate scaling equation for linear space charge tune shift, 
-        from Eq (1) in Hannes' and Isabelle's space charge report
-        at https://cds.cern.ch/record/2749453
-        (assuming lattice integral is constant and that this scaling 
-        stays constant, i.e. with  the same space charge tune shift 
-        - for now we ignore constants for a simpler expression
+        LEIR: Load Pb reference values and line, calculate present Pb space charge tune shift
+        Solve for maximum bunch intensity Nb_max with new particles
+
+        Parameters:
+        -----------
+        None
+
+        Returns:
+        --------
+        Nb_spaceChargeLimitLEIR : float
+            maximum bunch intensity from space charge limit
+        dQx, dQy : float
+            space charge tune shift related to this space charge limit
         """
-        # Specify if fully stripped ion or not
-        if fully_stripped:
-            charge = self.Z
-        else:
-            charge = self.Q
-        beta = self.beta(gamma)
-        return Nb*charge**2/(m *beta*gamma**2*epsilon*sigma_z)
-    
-    
-    def linearIntensityLimit(self, m, gamma, Nb_0, charge_0, m_0, gamma_0, fully_stripped=True):
-        """
-        Linear intensity limit for new ion species for given bunch intensity 
-        Nb_0 and parameters gamma_0, charge0, m_0 from reference ion species - assuming
-        that space charge stays constant, and that
-        emittance and bunch length are constant for all ion species
-        """
-        # Specify if fully stripped ion or not
-        if fully_stripped:
-            charge = self.Z
-        else:
-            charge = self.Q
-        beta_0 = self.beta(gamma_0)
-        beta = self.beta(gamma)
-        linearIntensityFactor = (m/m_0)*(charge_0/charge)**2*(beta/beta_0)*(gamma/gamma_0)**2   
+
+        # Generate new ion
+        LEIR_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_LEIR, gamma0 = self.LEIR_gamma_inj)
+
+        # Calculate max intensity from space charge tune shift, assuming we are at the maximum limit today
+        leir_sc = SC_Tune_Shifts()
+        Nb_spaceChargeLimitLEIR = leir_sc.maxIntensity_from_SC_integral(dQx_max=self.dQx0_LEIR_Pb, dQy_max=self.dQy0_LEIR_Pb,
+                                                                        twissTableXsuite=self.twiss_LEIR_Pb0,
+                                                                        particle_ref=LEIR_particle,
+                                                                        line_length=self.line_LEIR_length, 
+                                                                        beamParams=BeamParams_LEIR)
         
-        if self.debug_mode:
-            print(f"SPS intensity limit. Type: {self.ion_type}")
-            print("Fully stripped: {}".format(fully_stripped))
-            print("Q = {}, Z = {}".format(self.Q, self.Z))
-            print("Nb_0 = {:.2e}".format(Nb_0))
-            print("m = {:.2e} GeV, m0 = {:.2e} GeV".format(m, m_0))
-            print("charge = {:.1f}, charge_0 = {:.1f}".format(charge, charge_0))
-            print("beta = {:.5f}, beta_0 = {:.5f}".format(beta, beta_0))
-            print("gamma = {:.3f}, gamma_0 = {:.3f}".format(gamma, gamma_0))
-            print('Linear intensity factor: {:.3f}\n'.format(linearIntensityFactor))
-        return Nb_0*linearIntensityFactor 
-    
-    
-    def ion0_referenceValues(self):
-        """
-        Sets bunch intensity Nb, gamma factor and mass of reference ion species 
-        As of now, use reference values from Pb from Hannes 
-        For injection and extraction energy, use known Pb ion values from LIU report on 
-        https://edms.cern.ch/ui/file/1420286/2/LIU-Ions_beam_parameter_table.pdf
-        """
-        # LEIR
-        self.E_kin_per_A_LEIR_inj = 4.2e-3 # kinetic energy per nucleon in LEIR before RF capture, same for all species
-        self.E_kin_per_A_LEIR_extr = 7.22e-2 # kinetic energy per nucleon in LEIR at exit, same for all species
+        # Find tune shifts of space charge limits
+        dQx, dQy = leir_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_LEIR_Pb0,
+                                            particle_ref=LEIR_particle,
+                                            line_length=self.line_LEIR_length, 
+                                            Nb=Nb_spaceChargeLimitLEIR,
+                                            beamParams=BeamParams_LEIR)
         
-        # PS
-        self.E_kin_per_A_PS_inj = 7.22e-2 # GeV/nucleon according to LIU design report 
-        self.E_kin_per_A_PS_extr = 5.9 # GeV/nucleon according to LIU design report 
-        
-        # SPS
-        self.E_kin_per_A_SPS_inj = 5.9 # GeV/nucleon according to LIU design report 
-        self.E_kin_per_A_SPS_extr = 176.4 # GeV/nucleon according to LIU design report 
-        
-        # As of now, reference data exists only for Pb54+
-        if self.ion_type_ref == 'Pb':    
-            
-            # Pb ion values
-            self.m0_GeV = 193.687 # rest mass in GeV for Pb reference case 
-            self.Z0 = 82.0
-            
-            # LEIR - reference case for Pb54+ --> BEFORE stripping
-            self.Nq0_LEIR_extr = 10e10  # number of observed charges injected into LEIR
-            self.Q0_LEIR = 54.0
-            self.Nb0_LEIR_extr = self.Nq0_LEIR_extr/self.Q0_LEIR
-            self.gamma0_LEIR_inj = (self.m0_GeV + self.E_kin_per_A_LEIR_inj * 208)/self.m0_GeV
-            self.gamma0_LEIR_extr = (self.m0_GeV + self.E_kin_per_A_LEIR_extr * 208)/self.m0_GeV
-            
-            # PS - reference case for Pb54+ --> BEFORE stripping
-            self.Nq0_PS_extr =  6e10 # from November 2022 ionlifetime MD, previously 8e10  # number of observed charges injected at PS for nominal beam
-            self.Q0_PS = 54.0
-            self.Nb0_PS_extr = self.Nq0_PS_extr/self.Q0_PS
-            self.gamma0_PS_inj = (self.m0_GeV + self.E_kin_per_A_PS_inj * 208)/self.m0_GeV
-            self.gamma0_PS_extr = (self.m0_GeV + self.E_kin_per_A_PS_extr * 208)/self.m0_GeV
-            
-            # SPS - reference case for Pb82+ --> AFTER stripping
-            if not self.account_for_SPS_transmission:
-                self.SPS_transmission = 1.0
-            self.Nb0_SPS_extr = 2.21e8/self.SPS_transmission # outgoing ions per bunch from SPS (2015 values), adjusted for 62% transmission
-            self.Q0_SPS = 82.0
-            self.Nq0_SPS_extr = self.Nb0_SPS_extr*self.Q0_SPS
-            self.gamma0_SPS_inj = (self.m0_GeV + self.E_kin_per_A_SPS_inj * 208)/self.m0_GeV
-            self.gamma0_SPS_extr = (self.m0_GeV + self.E_kin_per_A_SPS_extr * 208)/self.m0_GeV
-    
-        else:
-            raise ValueError('Other reference ion type than Pb does not yet exist!')
-   
-    
-    def leir(self):
-        """
-        Calculate gamma at entrance and exit of the LEIR and transmitted bunch intensity 
-        using known Pb ion values from LIU report on 
-        https://edms.cern.ch/ui/file/1420286/2/LIU-Ions_beam_parameter_table.pdf
-        """
-        # Estimate gamma at extraction
-        if self.use_gammas_ref:
-            self.gamma_LEIR_inj = self.LEIR_gamma_inj_ref
-            self.gamma_LEIR_extr = self.LEIR_gamma_extr_ref
-        else: 
-            self.gamma_LEIR_inj = (self.mass_GeV + self.E_kin_per_A_LEIR_inj * self.A)/self.mass_GeV
-            
-            self.gamma_LEIR_extr =  np.sqrt(
-                                    1 + ((self.Q / 54) / (self.mass_GeV/self.m0_GeV))**2
-                                    * (self.gamma0_LEIR_extr**2 - 1)
-                                    )
-            
-            #self.gamma_LEIR_extr = (self.mass_GeV + self.E_kin_per_A_LEIR_extr * self.A)/self.mass_GeV         
-                
-        # Estimate number of charges at extraction - 10e10 charges for Pb54+, use this as scaling 
-        self.Nb_LEIR_extr = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_LEIR_inj,  
-                                               Nb_0 = self.Nb0_LEIR_extr, 
-                                               charge_0 = self.Q0_LEIR, # partially stripped charged state 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_LEIR_inj,  # use gamma at extraction
-                                               fully_stripped=False
-                                               )
-        
-        self.Nq_LEIR_extr = self.Nb_LEIR_extr*self.Q  # number of outgoing charges, before any stripping
+        return Nb_spaceChargeLimitLEIR, dQx, dQy
 
 
-    def stripper_foil_leir_ps(self):
+    def LEIR_tune_shifts(self, Nb_max=None):
         """
-        Stripper foil between LEIR and PS 
-        """
-        pass    
+        LEIR: Load Pb reference values and line, calculate space charge tune shift from
+        given bunch intensity
 
-    
-    def ps(self):
-        """
-        Calculate gamma at entrance and exit of the PS and transmitted bunch intensity 
-        """
-        # Estimate gamma at extraction
-        if self.use_gammas_ref:
-            self.gamma_PS_inj = self.PS_gamma_inj_ref
-            self.gamma_PS_extr = self.PS_gamma_extr_ref
-        else:         
+        Parameters:
+        -----------
+        Nb_max : float
+            maximum injected bunch intensity from previous accelerators, to calculate tune shift        
 
-            self.gamma_PS_inj =  self.gamma_LEIR_extr
-            self.gamma_PS_extr =  np.sqrt(
-                                    1 + (((self.Z if self.LEIR_PS_strip else self.Q) / 54) / (self.mass_GeV/self.m0_GeV))**2
-                                    * (self.gamma0_PS_extr**2 - 1)
-                                    )
+        Returns:
+        --------
+        dQx, dQy : float
+            space charge tune shift assuming maximum possible intensity from previous injectors
+        """
+        # If Nb_max is None, set to default value
+        if Nb_max is None:
+            Nb_max = BeamParams_LEIR.Nb
+
+        # Generate new ion
+        LEIR_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_LEIR, gamma0 = self.LEIR_gamma_inj)
+
+        # Find tune shifts of space charge limits
+        leir_sc = SC_Tune_Shifts()
+        dQx, dQy = leir_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_LEIR_Pb0,
+                                            particle_ref=LEIR_particle,
+                                            line_length=self.line_LEIR_length, 
+                                            Nb=Nb_max,
+                                            beamParams=BeamParams_LEIR)
         
-        # Estimate number of charges at extraction
-        self.Nb_PS_extr = self.linearIntensityLimit(
-                                                m = self.mass_GeV, 
-                                                gamma = self.gamma_PS_inj,  
-                                                Nb_0 = self.Nb0_PS_extr, 
-                                                charge_0 = self.Q0_PS, # partially stripped charged state 
-                                                m_0 = self.m0_GeV,  
-                                                gamma_0 = self.gamma0_PS_inj,  # use gamma at extraction,
-                                                fully_stripped=False
-                                                )
-        self.Nq_PS_extr = self.Nb_PS_extr*self.Q  # number of outgoing charges, before any stripping
-
-
-    def stripper_foil_ps_sps(self):
-        """
-        Stripper foil between PS and SPS 
-        """
-        pass   
-    
-    
-    def sps(self):
-        """
-        Calculate gamma at entrance and exit of the SPS, and transmitted bunch intensity 
-        Space charge limit comes from gamma at injection
-        """
-        # Calculate gamma at injection, simply scaling with the kinetic energy per nucleon as of today
-        if self.use_gammas_ref:
-            self.gamma_SPS_inj = self.SPS_gamma_inj_ref
-            self.gamma_SPS_extr = self.SPS_gamma_extr_ref
-        else:     
-            # In his notebook, Roderik considers same magnetic rigidity at PS extraction for all ion species: Brho = P/Q, P = m*gamma*beta*c
-            self.gamma_SPS_inj =  np.sqrt(
-                                    1 + (((self.Z if self.LEIR_PS_strip else self.Q) / 54) / (self.mass_GeV/self.m0_GeV))**2
-                                    * (self.gamma0_SPS_inj**2 - 1)
-                                )
-            self.gamma_SPS_extr = (self.mass_GeV + self.E_kin_per_A_SPS_extr * self.A)/self.mass_GeV
+        return dQx, dQy
         
-        print(f'SPS gamma: {self.gamma_SPS_inj}')
-        # Calculate outgoing intensity from linear scaling 
-        self.Nb_SPS_extr = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_SPS_inj,  
-                                               Nb_0 = self.Nb0_SPS_extr, # what we can successfully accelerate today
-                                               charge_0 = self.Q0_SPS, 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_SPS_inj,  # use gamma at extraction
-                                               fully_stripped=True
-                                               )
+        
+    def PS_SC_limit(self):
+        """
+        PS: Load Pb reference values and line, calculate present Pb space charge tune shift
+        Solve for maximum bunch intensity Nb_max with new particles
+
+        Parameters:
+        -----------
+        None
+
+        Returns:
+        --------
+        Nb_spaceChargeLimitLEIR : float
+            maximum possible injected bunch intensity from previous accelerators, to calculate new ion tune shift
+        dQx, dQy : float
+            space charge tune shift related to this space charge limit
+        """
+        # Generate new ion
+        PS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_PS, gamma0 = self.PS_gamma_inj)   
+        
+        # Calculate max intensity from space charge tune shift, assuming we are at the maximum limit today
+        ps_sc = SC_Tune_Shifts()
+        Nb_spaceChargeLimitPS = ps_sc.maxIntensity_from_SC_integral(dQx_max=self.dQx0_PS_Pb, dQy_max=self.dQy0_PS_Pb,
+                                                                    twissTableXsuite=self.twiss_PS_Pb0,
+                                                                    particle_ref=PS_particle,
+                                                                    line_length=self.line_PS_length, 
+                                                                    beamParams=BeamParams_PS)
+        
+        # Find tune shifts of space charge limit
+        dQx, dQy = ps_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_PS_Pb0,
+                                            particle_ref=PS_particle,
+                                            line_length=self.line_PS_length,
+                                            Nb=Nb_spaceChargeLimitPS,
+                                            beamParams=BeamParams_PS)
+        
+        return Nb_spaceChargeLimitPS, dQx, dQy
+
+
+    def PS_tune_shifts(self, Nb_max=None):
+        """
+        PS: Load Pb reference values and line, calculate space charge tune shift from
+        given bunch intensity
+
+        Parameters:
+        -----------
+        Nb_max : float
+            maximum injected bunch intensity from previous accelerators, to calculate tune shift        
+
+        Returns:
+        --------
+        dQx, dQy : float
+            space charge tune shift assuming maximum possible intensity from previous injectors
+        """
+        # If Nb_max is None, set to default value
+        if Nb_max is None:
+            Nb_max = BeamParams_PS.Nb
+
+        # Generate new ion
+        PS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_PS, gamma0 = self.PS_gamma_inj)   
+        
+        # Calculate max intensity from space charge tune shift, assuming we are at the maximum limit today
+        ps_sc = SC_Tune_Shifts()
+        dQx, dQy = ps_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_PS_Pb0,
+                                            particle_ref=PS_particle,
+                                            line_length=self.line_PS_length,
+                                            Nb=Nb_max,
+                                            beamParams=BeamParams_PS)
+        
+        return dQx, dQy
+
+
+    def SPS_SC_limit(self):
+        """
+        SPS: Load Pb reference values and line, calculate present Pb space charge tune shift
+        Solve for maximum bunch intensity Nb_max with new particles
+
+        Parameters:
+        -----------
+        None
+
+        Returns:
+        --------
+        Nb_spaceChargeLimitLEIR : float
+            maximum bunch intensity from space charge limit
+        dQx, dQy : float
+            space charge tune shift related to this space charge limit
+        """
+        # Generate new ion
+        SPS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_SPS, gamma0 = self.SPS_gamma_inj)
+
+        # Calculate max intensity from space charge tune shift, assuming we are at the maximum limit today
+        sps_sc = SC_Tune_Shifts()
+        Nb_spaceChargeLimitSPS = sps_sc.maxIntensity_from_SC_integral(dQx_max=self.dQx0_SPS_Pb, dQy_max=self.dQy0_SPS_Pb,
+                                                                      twissTableXsuite=self.twiss_SPS_Pb0,
+                                                                      particle_ref=SPS_particle,
+                                                                      line_length=self.line_SPS_length,
+                                                                      beamParams=BeamParams_SPS)
+        
+        dQx, dQy = sps_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_SPS_Pb0,
+                                                   particle_ref=SPS_particle,
+                                                   line_length=self.line_SPS_length,
+                                                   Nb=Nb_spaceChargeLimitSPS,
+                                                   beamParams=BeamParams_SPS)
+        
+        return Nb_spaceChargeLimitSPS, dQx, dQy
     
-        self.Nq_SPS_extr = self.Nb_SPS_extr*self.Z  # number of outgoing charges
-      
-    
-    def simulate_injection_SpaceCharge_limit(self):
-        """
-        Simulate space charge limit in full injection through Linac3, LEIR, PS and SPS for a given ion type
-        """
-        self.leir()
-        self.ps()
-        self.sps()
 
-
-    def simulate_SpaceCharge_intensity_limit_all_ions(self, return_dataframe=True):
+    def SPS_tune_shifts(self, Nb_max=None):
         """
-        Calculate intensity limits with linear space charge in 
-        Linac3, LEIR, PS and SPS for all ions given in table
+        SPS: Load Pb reference values and line, calculate space charge tune shift from
+        given bunch intensity
+
+        Parameters:
+        -----------
+        Nb_max : float
+            maximum injected bunch intensity from previous accelerators, to calculate tune shift        
+
+        Returns:
+        --------
+        dQx, dQy : float
+            space charge tune shift assuming maximum possible intensity from previous injectors
         """
-        
-        # Initiate row of ions per bunch (Nb) and charges per bunch (Nq)
-        self.ion_Nb_data = pd.DataFrame(index=self.full_ion_data.transpose().index)
-        self.ion_Nb_data["Nb_LEIR"] = np.NaN
-        self.ion_Nb_data["Nq_LEIR"] = np.NaN
-        self.ion_Nb_data["Nb_PS"] = np.NaN
-        self.ion_Nb_data["Nq_PS"] = np.NaN
-        self.ion_Nb_data["Nb_SPS"] = np.NaN
-        self.ion_Nb_data["Nq_SPS"] = np.NaN
-        
-        # Create dataframes or gamma injection and extraction data 
-        self.ion_gamma_inj_data = pd.DataFrame(columns=['LEIR', 'PS', 'SPS'], 
-                                               index=self.full_ion_data.transpose().index)
-        self.ion_gamma_extr_data = self.ion_gamma_inj_data.copy()
-        
-        # Iterate over all ions in data 
-        for i, ion_type in enumerate(self.full_ion_data.columns):
-            self.init_ion(ion_type)
-            self.simulate_injection_SpaceCharge_limit()
-            
-            # Add the intensities into a table 
-            self.ion_Nb_data["Nb_LEIR"][ion_type] = self.Nb_LEIR_extr
-            self.ion_Nb_data["Nq_LEIR"][ion_type] = self.Nq_LEIR_extr
-            self.ion_Nb_data["Nb_PS"][ion_type] = self.Nb_PS_extr
-            self.ion_Nb_data["Nq_PS"][ion_type] = self.Nq_PS_extr
-            self.ion_Nb_data["Nb_SPS"][ion_type] = self.Nb_SPS_extr
-            self.ion_Nb_data["Nq_SPS"][ion_type] = self.Nq_SPS_extr
+        # If Nb_max is None, set to default value
+        if Nb_max is None:
+            Nb_max = BeamParams_SPS.Nb
 
-            # Add the gamma of injection and extraction into a table 
-            self.ion_gamma_inj_data['LEIR'][ion_type] = self.gamma_LEIR_inj
-            self.ion_gamma_extr_data['LEIR'][ion_type] = self.gamma_LEIR_extr
-            self.ion_gamma_inj_data['PS'][ion_type] = self.gamma_PS_inj
-            self.ion_gamma_extr_data['PS'][ion_type] = self.gamma_PS_extr
-            self.ion_gamma_inj_data['SPS'][ion_type] = self.gamma_SPS_inj
-            self.ion_gamma_extr_data['SPS'][ion_type] = self.gamma_SPS_extr
+        # Generate new ion
+        SPS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_SPS, gamma0 = self.SPS_gamma_inj)
 
-        if return_dataframe:
-            return self.ion_Nb_data 
+        # Calculate max intensity from space charge tune shift, assuming we are at the maximum limit today
+        sps_sc = SC_Tune_Shifts()
+        dQx, dQy = sps_sc.calculate_SC_tuneshift(twissTableXsuite=self.twiss_SPS_Pb0,
+                                                   particle_ref=SPS_particle,
+                                                   line_length=self.line_SPS_length,
+                                                   Nb=Nb_max,
+                                                   beamParams=BeamParams_SPS)
+        return dQx, dQy
 
 
     def calculate_LHC_bunch_intensity(self):
         """
         Estimate LHC bunch intensity for a given ion species
-        through Linac3, LEIR, PS and SPS considering all the limits of the injectors
-        """
-        self.simulate_injection_SpaceCharge_limit()
-        
+        through Linac3, LEIR, PS and SPS considering full lattice integral space charge limits of the injectors
+        """        
         ### LINAC3 ### 
-        ionsPerPulseLinac3 = (self.linac3_current * self.linac3_pulseLength) / (self.Q * constants.e)
+        ionsPerPulseLinac3 = (self.linac3_current * self.linac3_pulseLength) / (self.Q_LEIR * constants.e)
         
         ### LEIR ###
-        spaceChargeLimitLEIR = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_LEIR_inj,  
-                                               Nb_0 = self.Nb0_LEIR_extr, 
-                                               charge_0 = self.Q0_LEIR, # partially stripped charged state 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_LEIR_inj,  # use gamma at LEIR extraction
-                                               fully_stripped=False
-                                               )
+        spaceChargeLimitLEIR, dQx0_LEIR, dQy0_LEIR = self.LEIR_SC_limit()  # space charge limit assuming same tune shifts as today
         
-        nPulsesLEIR = (min(7, math.ceil(spaceChargeLimitLEIR / (ionsPerPulseLinac3 * self.LEIR_injection_efficiency))) if self.nPulsesLEIR == 0 else self.nPulsesLEIR)
-        totalIntLEIR = ionsPerPulseLinac3 * nPulsesLEIR * self.LEIR_injection_efficiency
-        ionsPerBunchExtractedLEIR = self.LEIR_transmission * np.min([totalIntLEIR, spaceChargeLimitLEIR]) / self.LEIR_bunches
+        # Calculate maximum number of pulses (ceiling value) in LEIR due to space charge effects
+        if self.nPulsesLEIR is None:
+            #nPulsesLEIR_SC = min(Reference_Values.max_injections_into_LEIR, math.ceil(spaceChargeLimitLEIR / (ionsPerPulseLinac3 * Reference_Values.LEIR_injection_efficiency)))
+            nPulsesLEIR_default = Reference_Values.max_injections_into_LEIR
+        else:
+            nPulsesLEIR_default = self.nPulsesLEIR
+                              
+        # Calculate number of bunches to inject if we consider electron cooling
+        if self.account_for_LEIR_ecooling:
+            num_injections_LEIR_with_ecooling = math.ceil(Reference_Values.max_injections_into_LEIR / self.relative_ecooling_time_leir)
+            nPulsesLEIR = np.min([num_injections_LEIR_with_ecooling, nPulsesLEIR_default])
+        else:
+            nPulsesLEIR = nPulsesLEIR_default
+        
+        totalIntLEIR = ionsPerPulseLinac3 * nPulsesLEIR * Reference_Values.LEIR_injection_efficiency
+        ionsPerBunchExtractedLEIR = Reference_Values.LEIR_transmission * np.min([totalIntLEIR, spaceChargeLimitLEIR]) / self.LEIR_bunches
         LEIR_space_charge_limit_hit = True if totalIntLEIR > spaceChargeLimitLEIR else False 
+
+        dQx_LEIR, dQy_LEIR = self.LEIR_tune_shifts(Nb_max=totalIntLEIR) # calculate tune shifts for LEIR if max new intensity is injected
         
         #### PS ####
         ionsPerBunchInjectedPS = ionsPerBunchExtractedLEIR * (self.LEIR_PS_stripping_efficiency if self.LEIR_PS_strip else 1)
-        spaceChargeLimitPS = self.linearIntensityLimit(
-                                        m = self.mass_GeV, 
-                                        gamma = self.gamma_PS_inj,  
-                                        Nb_0 = self.Nb0_PS_extr, 
-                                        charge_0 = self.Q0_PS, # partially stripped charged state 
-                                        m_0 = self.m0_GeV,  
-                                        gamma_0 = self.gamma0_PS_inj,  # use gamma at PS inj
-                                        fully_stripped = self.LEIR_PS_strip # fully stripped if LEIR-PS strip
-                                        )
+        dQx_PS, dQy_PS = self.PS_tune_shifts(Nb_max=ionsPerBunchInjectedPS ) # calculate tune shifts for PS if max new intensity is injected
+        spaceChargeLimitPS, dQx0_PS, dQy0_PS = self.PS_SC_limit()
         
         # Check that injected momentum is not too low for the PS B-field
-        q_PS = self.Z if self.LEIR_PS_strip else self.Q
-        self.p_PS_inj = self.calcMomentum_from_gamma(self.gamma_PS_inj, q_PS)
-        self.Brho_PS_inj = self.calcBrho(self.p_PS_inj, q_PS) # same as LEIR extraction if no stripping, else will be different 
-        B_PS_inj = self.Brho_PS_inj / self.PS_rho
-        if B_PS_inj < self.PS_MinB:
+        self.p_PS_inj = self.calcMomentum_from_gamma(self.PS_gamma_inj, self.Q_PS)
+        self.Brho_PS_inj = self.calcBrho(self.p_PS_inj, self.Q_PS) # same as LEIR extraction if no stripping, else will be different 
+        B_PS_inj = self.Brho_PS_inj / Reference_Values.PS_rho
+        if B_PS_inj < Reference_Values.PS_MinB:
             self.PS_B_field_is_too_low = True
-        elif B_PS_inj > self.PS_MaxB:
-            print("\nA = {}, Q_low = {}, m_ion = {:.2f} u, Z = {}".format(self.A, self.Q_low, self.m_ion_in_u, self.Z))
+        elif B_PS_inj > Reference_Values.PS_MaxB:
+            print("\nA = {}, Q_PS = {}, m_ion = {:.2f} GeV, Z = {}".format(self.A, self.Q_PS, self.mass_GeV, self.Z))
             print('B = {:.4f} in PS at injection is too HIGH!'.format(B_PS_inj))
             raise ValueError("B field in PS is too high!")
         else:
             self.PS_B_field_is_too_low = False
         
-        # If space charge limit in PS is considered, choose the minimum between the SC limit and the extracted ionsPerBunchPS
-        if self.consider_PS_space_charge_limit:
-            ionsPerBunchPS = min(spaceChargeLimitPS, ionsPerBunchInjectedPS)
-            #if spaceChargeLimitPS < ionsPerBunchInjectedPS:
-            #    print("\nIon type: {}".format(self.ion_type))
-            #    print("Space charge limit PS: {:.3e} vs max injected ions per bunch PS: {:.3e}".format(spaceChargeLimitPS, ionsPerBunchInjectedPS))
-        else:
-            ionsPerBunchPS = ionsPerBunchInjectedPS 
+        # Select minimum between maxiumum possible injected intensity and PS space charge limit
+        ionsPerBunchPS = min(spaceChargeLimitPS, ionsPerBunchInjectedPS)
         PS_space_charge_limit_hit = True if ionsPerBunchInjectedPS > spaceChargeLimitPS else False 
-        ionsPerBunchExtracted_PS = ionsPerBunchPS * self.PS_transmission / self.PS_splitting # maximum intensity without SC
+        ionsPerBunchExtracted_PS = ionsPerBunchPS * Reference_Values.PS_transmission / self.PS_splitting # maximum intensity without SC
         
         # Calculate ion transmission for SPS 
-        ionsPerBunchSPSinj = ionsPerBunchExtracted_PS * (self.PS_SPS_transmission_efficiency if self.Z == self.Q or self.LEIR_PS_strip else self.PS_SPS_stripping_efficiency)
-        spaceChargeLimitSPS = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_SPS_inj,  
-                                               Nb_0 = self.Nb0_SPS_extr, 
-                                               charge_0 = self.Q0_SPS, 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_SPS_inj, # use gamma at SPS inj
-                                               fully_stripped=True
-                                               )
+        ionsPerBunchSPSinj = ionsPerBunchExtracted_PS * (Reference_Values.PS_SPS_transmission_efficiency if self.LEIR_PS_strip else Reference_Values.PS_SPS_stripping_efficiency)
+        dQx_SPS, dQy_SPS = self.SPS_tune_shifts(Nb_max=ionsPerBunchSPSinj) # calculate tune shifts for SPS if max new intensity is injected
+        spaceChargeLimitSPS, dQx0_SPS, dQy0_SPS = self.SPS_SC_limit()
         SPS_space_charge_limit_hit = True if ionsPerBunchSPSinj > spaceChargeLimitSPS else False
-        ionsPerBunchLHC = min(spaceChargeLimitSPS, ionsPerBunchSPSinj) * self.SPS_transmission * self.SPS_slipstacking_transmission
+        ionsPerBunchLHC = min(spaceChargeLimitSPS, ionsPerBunchSPSinj) * Reference_Values.SPS_transmission * Reference_Values.SPS_slipstacking_transmission
 
         result = {
             "Ion": self.ion_type,
-            "chargeBeforeStrip": int(self.Q),
             "atomicNumber": int(self.Z),
             "massNumber": int(self.A),
+            "Q_LEIR": int(self.Q_LEIR),
+            "Q_PS": int(self.Q_PS),
+            "Q_SPS": int(self.Q_SPS),
             "Linac3_current [A]": self.linac3_current,
             "Linac3_pulse_length [s]": self.linac3_pulseLength, 
             "LEIR_numberofPulses": nPulsesLEIR,
-            "LEIR_injection_efficiency": self.LEIR_injection_efficiency, 
+            "LEIR_injection_efficiency": Reference_Values.LEIR_injection_efficiency, 
             "LEIR_splitting": self.LEIR_bunches,
-            "LEIR_transmission": self.LEIR_transmission, 
+            "LEIR_transmission": Reference_Values.LEIR_transmission, 
             "PS_splitting": self.PS_splitting, 
-            "PS_transmission": self.PS_transmission, 
-            "PS_SPS_stripping_efficiency": self.PS_SPS_stripping_efficiency, 
-            "SPS_transmission": self.SPS_transmission, 
+            "PS_transmission": Reference_Values.PS_transmission, 
+            "PS_SPS_stripping_efficiency": Reference_Values.PS_SPS_stripping_efficiency, 
+            "SPS_transmission": Reference_Values.SPS_transmission, 
             "Linac3_ionsPerPulse": ionsPerPulseLinac3,
             "LEIR_maxIntensity": totalIntLEIR,
             "LEIR_space_charge_limit": spaceChargeLimitLEIR,
@@ -520,36 +512,135 @@ class InjectorChain:
             "SPS_space_charge_limit": spaceChargeLimitSPS,
             "LHC_ionsPerBunch": ionsPerBunchLHC,
             "LHC_chargesPerBunch": ionsPerBunchLHC * self.Z,
-            "LEIR_gamma_inj": self.gamma_LEIR_inj,
-            "LEIR_gamma_extr": self.gamma_LEIR_extr,
-            "PS_gamma_inj": self.gamma_PS_inj,
-            "PS_gamma_extr": self.gamma_PS_extr,
-            "SPS_gamma_inj": self.gamma_SPS_inj,
-            "SPS_gamma_extr": self.gamma_SPS_extr,
+            "LEIR_gamma_inj": self.LEIR_gamma_inj,
+            "LEIR_gamma_extr": self.LEIR_gamma_extr,
+            "PS_gamma_inj": self.PS_gamma_inj,
+            "PS_gamma_extr": self.PS_gamma_extr,
+            "SPS_gamma_inj": self.SPS_gamma_inj,
+            "SPS_gamma_extr": self.SPS_gamma_extr,
             "PS_B_field_is_too_low": self.PS_B_field_is_too_low,
             "LEIR_space_charge_limit_hit": LEIR_space_charge_limit_hit,
-            "consider_PS_space_charge_limit": self.consider_PS_space_charge_limit,
             "PS_space_charge_limit_hit": PS_space_charge_limit_hit,
             "SPS_space_charge_limit_hit": SPS_space_charge_limit_hit,
             "LEIR_ratio_SC_limit_maxIntensity": spaceChargeLimitLEIR / totalIntLEIR,
             "PS_ratio_SC_limit_maxIntensity": spaceChargeLimitPS / ionsPerBunchInjectedPS,
-            "SPS_ratio_SC_limit_maxIntensity": spaceChargeLimitSPS / ionsPerBunchSPSinj
+            "SPS_ratio_SC_limit_maxIntensity": spaceChargeLimitSPS / ionsPerBunchSPSinj,
+            "dQx LEIR SC limit tune shift": dQx0_LEIR,
+            "dQy LEIR SC limit tune shift": dQy0_LEIR,
+            "dQx PS SC limit tune shift": dQx0_PS,
+            "dQy PS SC limit tune shift": dQy0_PS,
+            "dQx SPS SC limit tune shift": dQx0_SPS,
+            "dQy SPS SC limit tune shift": dQy0_SPS,
+            "dQx LEIR tune shift with Nb_max": dQx_LEIR,
+            "dQy LEIR tune shift with Nb_max": dQy_LEIR,
+            "dQx PS tune shift with Nb_max": dQx_PS,
+            "dQy PS tune shift with Nb_max": dQy_PS,
+            "dQx SPS tune shift with Nb_max": dQx_SPS,
+            "dQy SPS tune shift with Nb_max": dQy_SPS
         }
 
         # Add key of LEIR-PS stripping efficiency if this is done 
         if self.LEIR_PS_strip:
             result["LEIR_PS_strippingEfficiency"] = self.LEIR_PS_stripping_efficiency
+            
+        if self.account_for_LEIR_ecooling:
+            result["LEIR_relative_ecooling_time"] = self.relative_ecooling_time_leir
+            result["LEIR_numberofPulses_ecooling"] = num_injections_LEIR_with_ecooling
     
         return result
 
+        
+    def calculate_IBS_growth_rates(self):
+        """
+        Calculate IBS growth rates for initiated ion species and beam parameters
 
-    def calculate_LHC_bunch_intensity_all_ion_species(self, save_csv=False, output_name='output'):
+        Returns:
+        --------
+        self.ion_str : str
+            ion type in string format
+        growth_rates_dict : dict
+            IBS Nagaitsev growth rates dictionary
+        """
+
+        # Instantiate IBS analytical model
+        IBS = IBS_Growth_Rates()
+        
+        # LEIR growth rates
+        leir_line = self.line_LEIR_Pb0.copy()
+        leir_line.particle_ref = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_LEIR, gamma0 = self.LEIR_gamma_inj)
+        growth_rates_leir = IBS.get_growth_rates(leir_line, BeamParams_LEIR)
+        
+        # PS growth rates
+        ps_line = self.line_PS_Pb0.copy()
+        ps_line.particle_ref = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_PS, gamma0 = self.PS_gamma_inj)
+        growth_rates_ps = IBS.get_growth_rates(ps_line, BeamParams_PS)
+        
+        # SPS growth rates
+        sps_line = self.line_SPS_Pb0.copy()
+        sps_line.particle_ref = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_SPS, gamma0 = self.SPS_gamma_inj)
+        growth_rates_sps = IBS.get_growth_rates(sps_line, BeamParams_SPS)
+        print('IBS growth rates calculated for {}'.format(self.ion_str))
+
+        growth_rates_dict = {'LEIR Tx': growth_rates_leir[0],
+                             'LEIR Ty': growth_rates_leir[1],
+                             'LEIR Tz': growth_rates_leir[2],
+                             'PS Tx': growth_rates_ps[0],
+                             'PS Ty': growth_rates_ps[1],
+                             'PS Tz': growth_rates_ps[2],
+                             'SPS Tx': growth_rates_sps[0],
+                             'SPS Ty': growth_rates_sps[1], 
+                             'SPS Tz': growth_rates_sps[2]}
+        
+        return self.ion_str, growth_rates_dict
+
+
+    def calculate_IBS_growth_rates_all_ion_species(self, save_csv=True, output_name='IBS_growth_rates.csv'):
+        """
+        Calculate analytical Nagaitsev IBS growth rates for all ions in LEIR, PS and SPS
+
+        Parameters
+        ----------
+        save_csv : bool
+            whether to save output in csv format, else return it
+        output_name : str
+            name of csv file to be generated
+
+        Returns
+        -------
+        df_IBS : pd.DataFrame
+            dataframe containing all IBS growth rates
+        """        
+        # Check that output directory exists
+        os.makedirs('output_csv', exist_ok=True)
+
+        # Iterate over all ions in data 
+        all_ion_IBS_dict = {}
+        for ion_type in self.full_ion_data.columns:
+            
+            # Initiate the correct ion and calculate growth rates
+            self.init_ion(ion_type)
+            ion_str, growth_rates_dict = self.calculate_IBS_growth_rates()
+            all_ion_IBS_dict[ion_str] = growth_rates_dict
+
+        df_IBS = pd.DataFrame(all_ion_IBS_dict)
+        if save_csv:
+            df_IBS.to_csv('output_csv/{}'.format(output_name))
+        else:
+            return df_IBS
+
+
+    def calculate_LHC_bunch_intensity_all_ion_species(self, save_csv=True, output_name='output'):
         """
         Estimate LHC bunch intensity for all ion species provided in table
         through Linac3, LEIR, PS and SPS considering all the limits of the injectors
+
+        Returns:
+        --------
+        df_all_ions : pd.DataFrame
+            pandas dataframe containing result of all ions
         """
         # Check that output directory exists
-        os.makedirs(self.save_path, exist_ok=True)
+        os.makedirs('output_csv', exist_ok=True)
         
         # Initialize full dicionary
         full_result = defaultdict(list)
@@ -572,80 +663,18 @@ class InjectorChain:
         
         # Save CSV file if desired 
         if save_csv:
-            float_columns = df_all_ions.select_dtypes(include=['float']).columns
+            
+            # First save full CSV
             df_save = df_all_ions.copy()
-            df_save[float_columns]  = df_save[float_columns].applymap(self.format_large_numbers)
-            df_save = df_save.T
-            df_save.to_csv("{}/{}.csv".format(self.save_path, output_name))
+            df_save0 = df_save.T
+            df_save0.to_csv("output_csv/{}.csv".format(output_name))
+            
+            # Then save copy in exponential form with decimal number for paper - only some columns
+            float_columns = df_all_ions.select_dtypes(include=['float']).columns
+            for col in float_columns:
+                df_save[col] = df_save[col].apply(lambda x: '{:.1e}'.format(x))
+            df_SC_and_max_intensity = df_save[['LEIR_maxIntensity', 'LEIR_space_charge_limit', 'PS_maxIntensity', 'PS_space_charge_limit', 
+                        'SPS_maxIntensity', 'SPS_space_charge_limit', 'LHC_ionsPerBunch', 'LHC_chargesPerBunch']]
+            df_SC_and_max_intensity.to_csv("output_csv/{}_for_paper.csv".format(output_name), index=True)
             
         return df_all_ions
-    
-    
-    def space_charge_limit_effect_on_LHC_bunch_intensity(self):
-        """
-        Propagate estimated bunch intensity into LHC for a given ion species
-        considering space charge limit only in LEIR, PS or SPS 
-        --> considering first default baseline case
-        """
-        self.simulate_injection_SpaceCharge_limit()
-    
-        #### LEIR SPACE CHARGE LIMIT ####
-        spaceChargeLimitLEIR = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_LEIR_inj,  
-                                               Nb_0 = self.Nb0_LEIR_extr, 
-                                               charge_0 = self.Q0_LEIR, # partially stripped charged state 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_LEIR_inj,  # use gamma at LEIR extraction
-                                               fully_stripped=False
-                                               )
-        
-        ionsPerBunchLHC_LEIR_lim = spaceChargeLimitLEIR * self.LEIR_transmission \
-                                *  (self.LEIR_PS_stripping_efficiency if self.LEIR_PS_strip else 1) \
-                                * self.PS_transmission * (self.PS_SPS_transmission_efficiency if self.Z == self.Q or self.LEIR_PS_strip else self.PS_SPS_stripping_efficiency) \
-                                * self.SPS_transmission / (self.LEIR_bunches * self.PS_splitting)  # divide by number of pulses into LEIR and PS splitting
-        
-
-        #### PS SPACE CHARGE LIMIT ####
-        spaceChargeLimitPS = self.linearIntensityLimit(
-                                        m = self.mass_GeV, 
-                                        gamma = self.gamma_PS_inj,  
-                                        Nb_0 = self.Nb0_PS_extr, 
-                                        charge_0 = self.Q0_PS, # partially stripped charged state 
-                                        m_0 = self.m0_GeV,  
-                                        gamma_0 = self.gamma0_PS_inj,  # use gamma at PS inj
-                                        fully_stripped = self.LEIR_PS_strip # fully stripped if LEIR-PS strip
-                                        )
-
-        ionsPerBunchLHC_PS_lim = spaceChargeLimitPS * self.PS_transmission \
-                                * (self.PS_SPS_transmission_efficiency if self.Z == self.Q or self.LEIR_PS_strip else self.PS_SPS_stripping_efficiency) \
-                                * self.SPS_transmission / self.PS_splitting
-
-
-        #### SPS SPACE CHARGE LIMIT ####
-        spaceChargeLimitSPS = self.linearIntensityLimit(
-                                               m = self.mass_GeV, 
-                                               gamma = self.gamma_SPS_inj,  
-                                               Nb_0 = self.Nb0_SPS_extr, 
-                                               charge_0 = self.Q0_SPS, 
-                                               m_0 = self.m0_GeV,  
-                                               gamma_0 = self.gamma0_SPS_inj, # use gamma at SPS inj
-                                               fully_stripped=True
-                                               )
-        ionsPerBunchLHC_SPS_lim = spaceChargeLimitSPS * self.SPS_transmission * self.SPS_slipstacking_transmission        
-        
-        # Array of Nb injected into LHC considering SC limits in LEIR, PS and SPS
-        Nb_SC_limits = np.array([ionsPerBunchLHC_LEIR_lim, ionsPerBunchLHC_PS_lim, ionsPerBunchLHC_SPS_lim])
-
-        return Nb_SC_limits
-        
-    
-    
-    
-    def format_large_numbers(self, x):
-        """
-        Converts large or small floats to exponential form 
-        """
-        if abs(x) >= 1e6 or abs(x) < 1e-3:
-            return f'{x:.5e}'
-        return x
